Started refactoring by putting there the functions in a controller
