+++ main.go
@@ -2,104 +2,18 @@ package main
 
 import (
 	"github.com/hydeenoble/mux-rest-api/app"
-	"encoding/json"
-	"net/http"
-	"math/rand"
-	"strconv"
-	"github.com/gorilla/mux"
 )
-
-// Book Struct (Model)
-type Book struct {
-	ID string `json:"id"`
-	Isbn string `json:"isbn"`
-	Title string `json:"title"`
-	Author *Author `json:"author"`
-}
-
-// Author Struct
-type Author struct {
-	Firstname string `json:"firstname"`
-	Lastname string `json:"lastname"`
-}
-
-// Init books var as a slice Book Struct
-
-var books []Book
-
-// Get All Books
-func getBooks(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(books)
-}
-
-// Get single Book
-func getBook(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	params := mux.Vars(r) // Get Params 
-	// Loop through books and find with ID 
-	for _, book := range books {
-		if book.ID == params["id"] {
-			json.NewEncoder(w).Encode(book)
-			return
-		}
-	}
-	json.NewEncoder(w).Encode(&Book{})
-}
-
-//Create a New Book
-func createBook(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	var book Book
-	_ = json.NewDecoder(r.Body).Decode(&book)
-	book.ID = strconv.Itoa(rand.Intn(10000000))
-	books = append(books, book)
-	json.NewEncoder(w).Encode(book)
-}
-
-// Update a Book
-func updateBook(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	params := mux.Vars(r)
-	for index, item := range books {
-		if item.ID == params["id"] {
-			books = append(books[:index], books[index+1:]...)
-			var book Book
-			_ = json.NewDecoder(r.Body).Decode(&book)
-			book.ID = strconv.Itoa(rand.Intn(10000000))
-			books = append(books, book)
-			json.NewEncoder(w).Encode(book)
-			return
-		}
-	}
-	json.NewEncoder(w).Encode(books)
-}
-// Delete a Book
-func deleteBook(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	params := mux.Vars(r)
-	for index, item := range books {
-		if item.ID == params["id"] {
-			books = append(books[:index], books[index+1:]...)
-			break
-		}
-	}
-	json.NewEncoder(w).Encode(books)
-}
 
 func main() {
 	// testing external function
 	app := &app.App{}
+	app.Initialize()
 	// Init Router
-	r := mux.NewRouter()
+	// r := mux.NewRouter()
 
 	// Mock Data - @todo - implement DB
 
-	books = append(books, Book{ID: "1", Isbn: "448743", Title: "Book One", Author: &Author{Firstname: "John", Lastname: "Doe"}})
-	books = append(books, Book{ID: "2", Isbn: "433323", Title: "Book Two", Author: &Author{Firstname: "Steve", Lastname: "Smith"}})
-
 	
-
 	// log.Fatal(http.ListenAndServe(":8000", r))
 	app.Run(8000)
